feat: new test script that outputs to NeoPixel
Problem: the colour displayed on the neopixel is the inverse to the
colour detected.
This is because what is measured is the duration of the
pulse using pulseIn(). An object reflects its own colour into the sensor
at a much higher frequency than other colours in the light spectrum.
Since the duration of a signal is 1/frequency, lower values read by
the sensor actually mean a higher intensity of colour is being read.
Todo: Find a way to translate colour readings into a useful format
for the NeoPixel LEDs to display accurate colours.